requirements.txt
Django==3.2.9
requests~=2.27.1
django-allauth~=0.49.0